% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaseControl.R
\docType{package}
\name{CaseControl-package}
\alias{CaseControl}
\alias{CaseControl-package}
\title{CaseControl: Case-Control}
\description{
CaseControl is an R package for performing (nested) matched case-control analyses in an observational database in the OMOP Common Data Model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/CaseControl}
  \item Report bugs at \url{https://github.com/OHDSI/CaseControl/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
