% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{computeMdrr}
\alias{computeMdrr}
\title{Compute the minimum detectable relative risk}
\usage{
computeMdrr(caseControlData, alpha = 0.05, power = 0.8, twoSided = TRUE)
}
\arguments{
\item{caseControlData}{A data frame describing the cases and controls as created using the
\code{\link{createCaseControlData}} function. This should at least have these
columns: isCase, exposed.}

\item{alpha}{Type I error.}

\item{power}{1 - beta, where beta is the type II error.}

\item{twoSided}{Consider a two-sided test?}
}
\value{
A data frame with the MDRR and some counts.
}
\description{
Compute the minimum detectable relative risk
}
\details{
Compute the minimum detectable relative risk (MDRR) for a given study population, using the actual
observed sample size and number of exposed controls. Computations by  Miettinnen (1969) and
Rothman and Boice (1979) are used. Based on and verified using Ken Rothman's EpiSheet.
}
\references{
Miettinen OS (1969) Individual matching in the case of all or none responses. Biometrics, 25, 339-354.

Rothman KJ, Boice JD (1979) Epidemiologic Analysis with a Programmable Calculator. NIH Publication No.79-1649.
}
