% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{createCcAnalysis}
\alias{createCcAnalysis}
\title{Create a case-control analysis specification}
\usage{
createCcAnalysis(analysisId = 1, description = "", exposureType = NULL,
  outcomeType = NULL, nestingCohortType = NULL, getDbCaseDataArgs,
  selectControlsArgs, getDbExposureDataArgs, createCaseControlDataArgs,
  fitCaseControlModelArgs)
}
\arguments{
\item{analysisId}{An integer that will be used later to refer to this specific set
of analysis choices.}

\item{description}{A short description of the analysis.}

\item{exposureType}{If more than one exposure is provided for each
exposureOutcomeNestingCohort, this field should be used to select
the specific exposure to use in this analysis.}

\item{outcomeType}{If more than one outcome is provided for each
exposureOutcomeNestingCohort, this field should be used to select
the specific outcome to use in this analysis.}

\item{nestingCohortType}{If more than one nesting cohort is provided for each
exposureOutcomeNestingCohort, this field should be used to select
the specific nesting cohort to use in this analysis.}

\item{getDbCaseDataArgs}{An object representing the arguments to be used when calling the
\code{\link{createGetDbCaseDataArgs}} function.}

\item{selectControlsArgs}{An object representing the arguments to be used when calling the
\code{\link{createSelectControlsArgs}} function.}

\item{getDbExposureDataArgs}{An object representing the arguments to be used when calling the
\code{\link{createGetDbExposureDataArgs}} function.}

\item{createCaseControlDataArgs}{An object representing the arguments to be used when calling the
\code{\link{createCreateCaseControlDataArgs}} function.}

\item{fitCaseControlModelArgs}{An object representing the arguments to be used when calling the
\code{\link{createFitCaseControlModelArgs}} function.}
}
\description{
Create a case-control analysis specification
}
\details{
Create a set of analysis choices, to be used with the \code{\link{runCcAnalyses}} function.
}

