% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createCreateCaseControlDataArgs}
\alias{createCreateCaseControlDataArgs}
\title{Create a parameter object for the function createCaseControlData}
\usage{
createCreateCaseControlDataArgs(firstExposureOnly = FALSE,
  riskWindowStart = 0, riskWindowEnd = 0, exposureWashoutPeriod = 0)
}
\arguments{
\item{firstExposureOnly}{Should only the first exposure per subject be included?}

\item{riskWindowStart}{The start of the risk window (in days) relative to the index
date.This number should be non-positive.}

\item{riskWindowEnd}{The end of the risk window (in days) relative to the index date.
Thisnumber should be non-positive.}

\item{exposureWashoutPeriod}{Minimum required numbers of days of observation for inclusion of an
exposure.}
}
\description{
Create a parameter object for the function createCaseControlData
}
\details{
Create an object defining the parameter values.
}
