% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{createExposureOutcomeNestingCohort}
\alias{createExposureOutcomeNestingCohort}
\title{Create exposure-outcome-nesting-cohort combinations.}
\usage{
createExposureOutcomeNestingCohort(exposureId, outcomeId,
  nestingCohortId = NULL)
}
\arguments{
\item{exposureId}{A concept ID indentifying the target drug in the exposure table. If multiple
strategies for picking the exposure will be tested in the analysis, a named list
of numbers can be provided instead. In the analysis, the name of the number to
be used can be specified using the #' \code{exposureType} parameter in the
\code{\link{createCcAnalysis}} function.}

\item{outcomeId}{A concept ID indentifying the outcome in the outcome table. If multiple
strategies for picking the outcome will be tested in the analysis, a named list
of numbers can be provided instead. In the analysis, the name of the number to
be used can be specified using the \code{outcomeType} parameter in the
\code{\link{createCcAnalysis}} function.}

\item{nestingCohortId}{A concept ID indentifying the nesting cohort in the nesting cohort table. If multiple
strategies for picking the nesting cohort will be tested in the analysis, a named list
of numbers can be provided instead. In the analysis, the name of the number to
be used can be specified using the \code{nestingCohortType} parameter in the
\code{\link{createCcAnalysis}} function.}
}
\description{
Create exposure-outcome-nesting-cohort combinations.
}
\details{
Create a set of hypotheses of interest, to be used with the \code{\link{runCcAnalyses}} function.
}

