% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createFitCaseControlModelArgs}
\alias{createFitCaseControlModelArgs}
\title{Create a parameter object for the function fitCaseControlModel}
\usage{
createFitCaseControlModelArgs(
  useCovariates = FALSE,
  excludeCovariateIds = c(),
  includeCovariateIds = c(),
  prior = createPrior("laplace", useCrossValidation = TRUE),
  control = createControl(cvType = "auto", startingVariance = 0.01, tolerance = 2e-07,
    cvRepetitions = 10, selectorType = "byPid", noiseLevel = "quiet")
)
}
\arguments{
\item{useCovariates}{Whether to use the covariates in the caseControlsExposure.}

\item{excludeCovariateIds}{Exclude these covariates from the model.}

\item{includeCovariateIds}{Include only these covariates in the model.}

\item{prior}{The prior used to fit the model. See createPrior for details.}

\item{control}{The control object used to control the cross-validation used to
determine the hyperparameters of the prior (if applicable). See
createControl for details.}
}
\description{
Create a parameter object for the function fitCaseControlModel
}
\details{
Create an object defining the parameter values.
}
