% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createGetDbCaseDataArgs}
\alias{createGetDbCaseDataArgs}
\title{Create a parameter object for the function getDbCaseData}
\usage{
createGetDbCaseDataArgs(
  useNestingCohort = FALSE,
  useObservationEndAsNestingEndDate = TRUE,
  getVisits = FALSE,
  studyStartDate = "",
  studyEndDate = "",
  maxNestingCohortSize = 1e+07,
  maxCasesPerOutcome = 5e+05
)
}
\arguments{
\item{useNestingCohort}{Should the study be nested in a cohort (e.g. people with
a specific indication)? If not, the study will be nested
in the general population.}

\item{useObservationEndAsNestingEndDate}{When using a nesting cohort, should the observation
period end date be used instead of the cohort end date?}

\item{getVisits}{Get data on visits? This is needed when matching on visit
date is requested later on.}

\item{studyStartDate}{A calendar date specifying the minimum date where data is
used. Date format is 'yyyymmdd'.}

\item{studyEndDate}{A calendar date specifying the maximum date where data is
used. Date format is 'yyyymmdd'.}

\item{maxNestingCohortSize}{If the nesting cohort is larger than this number it will
be sampled to this size. maxCohortSize = 0 indicates no
maximum size.}

\item{maxCasesPerOutcome}{If there are more than this number of cases for a single
outcome cases will be sampled to this size.
maxCasesPerOutcome = 0 indicates no maximum size.}
}
\description{
Create a parameter object for the function getDbCaseData
}
\details{
Create an object defining the parameter values.
}
