% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createGetDbExposureDataArgs}
\alias{createGetDbExposureDataArgs}
\title{Create a parameter object for the function getDbExposureData}
\usage{
createGetDbExposureDataArgs(covariateSettings = NULL)
}
\arguments{
\item{covariateSettings}{An object of type covariateSettings as created using the
createCovariateSettings function in the FeatureExtraction package. If
NULL then no covariate data is retrieved.}
}
\description{
Create a parameter object for the function getDbExposureData
}
\details{
Create an object defining the parameter values.
}
