% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectControls.R
\name{createSamplingCriteria}
\alias{createSamplingCriteria}
\title{Create sampling criteria}
\usage{
createSamplingCriteria(controlsPerCase = 1, seed = 1)
}
\arguments{
\item{controlsPerCase}{Maximum number of controls to select per case.}

\item{seed}{The number generator seed. A null value sets seed via \code{\link{Sys.time}}.}
}
\value{
A settings object to be used in the \code{\link{selectControls}} function.
}
\description{
Criteria to use when controls are simply sampled from the (nesting) population.
}
