% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExposureData.R
\name{createSimpleCovariateSettings}
\alias{createSimpleCovariateSettings}
\title{Create simple covariate settings}
\usage{
createSimpleCovariateSettings(
  useDemographicsAgeGroup = TRUE,
  useDemographicsGender = TRUE
)
}
\arguments{
\item{useDemographicsAgeGroup}{Age of the subject on the index date (in 5 year age groups).}

\item{useDemographicsGender}{Gender of the subject.}
}
\value{
An object of type \code{SimpleCovariateSettings}
}
\description{
Create simple covariate settings
}
\details{
Create simple settings for covariates that can be constructed using only the
data in the \code{CaseData} object.
}
