% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFitting.R
\name{fitCaseControlModel}
\alias{fitCaseControlModel}
\title{Fit the case-control model}
\usage{
fitCaseControlModel(caseControlData)
}
\arguments{
\item{caseControlData}{A data frame as generated by the \code{\link{createCaseControlData}} function.}
}
\value{
An object of type \code{outcomeModel}.
}
\description{
Fit the case-control model
}
\details{
Fits the model using a conditional logistic regression.
}

