% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectControls.R
\name{getAttritionTable}
\alias{getAttritionTable}
\title{Get the attrition table for a population}
\usage{
getAttritionTable(caseControls)
}
\arguments{
\item{caseControls}{A data frame of cases and controls as generated by the function
\code{\link{selectControls}}.}
}
\value{
A data frame specifying the number of cases and events after various steps of filtering.
}
\description{
Get the attrition table for a population
}
