% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExposureData.R
\name{getDbExposureData}
\alias{getDbExposureData}
\title{Get exposure data for cases and controls from a database}
\usage{
getDbExposureData(connectionDetails, caseControls, oracleTempSchema = NULL,
  exposureDatabaseSchema, exposureTable = "drug_era", exposureIds = c())
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{caseControls}{A data frame as generated by the \code{\link{selectControls}} function.}
}
\description{
Get exposure data for cases and controls from a database
}

