% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaseData.R
\name{isCaseData}
\alias{isCaseData}
\title{Check whether an object is a CaseData object}
\usage{
isCaseData(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A logical value.
}
\description{
Check whether an object is a CaseData object
}
