% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExposureData.R
\name{loadCaseControlsExposure}
\alias{loadCaseControlsExposure}
\title{Load the caseControlsExposure data from a folder}
\usage{
loadCaseControlsExposure(folder, readOnly = TRUE)
}
\arguments{
\item{folder}{The name of the folder containing the data.}

\item{readOnly}{If true, the data is opened read only.}
}
\value{
An object of class \code{caseControlsExposure}.
}
\description{
\code{loadCaseControlsExposure} loads an object of type caseControlsExposure from a folder in the
file system.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}

