% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaseData.R
\name{loadCaseData}
\alias{loadCaseData}
\title{Load the covariate data from a folder}
\usage{
loadCaseData(file)
}
\arguments{
\item{file}{The name of the file containing the data.}
}
\value{
An object of class \code{CaseData}.
}
\description{
\code{loadCaseData} loads an object of type caseData from a file in the file system.
}
\details{
The data will be read from the file specified by the user.
}
