% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{loadCcAnalysisList}
\alias{loadCcAnalysisList}
\title{Load a list of ccAnalysis from file}
\usage{
loadCcAnalysisList(file)
}
\arguments{
\item{file}{The name of the file}
}
\value{
A list of objects of type \code{ccAnalysis}.
}
\description{
Load a list of objects of type \code{ccAnalysis} from file. The file is in JSON format.
}

