% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{loadExposureOutcomeNestingCohortList}
\alias{loadExposureOutcomeNestingCohortList}
\title{Load a list of exposureOutcomeNestingCohort from file}
\usage{
loadExposureOutcomeNestingCohortList(file)
}
\arguments{
\item{file}{The name of the file}
}
\value{
A list of objects of type \code{drugComparatorOutcome}.
}
\description{
Load a list of objects of type \code{exposureOutcomeNestingCohort} from file. The file is in JSON
format.
}
