% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{runCcAnalyses}
\alias{runCcAnalyses}
\title{Run a list of analyses}
\usage{
runCcAnalyses(connectionDetails, cdmDatabaseSchema,
  oracleTempSchema = cdmDatabaseSchema,
  exposureDatabaseSchema = cdmDatabaseSchema,
  exposureTable = "drug_era",
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_era",
  nestingCohortDatabaseSchema = cdmDatabaseSchema,
  nestingCohortTable = "condition_era", outputFolder = "./CcOutput",
  ccAnalysisList, exposureOutcomeNestingCohortList,
  prefetchExposureData = FALSE, compressCaseDataFiles = FALSE,
  getDbCaseDataThreads = 1, selectControlsThreads = 1,
  getDbExposureDataThreads = 1, createCaseControlDataThreads = 1,
  fitCaseControlModelThreads = 1, cvThreads = 1)
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using
the function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On
SQL Server, this should specifiy both the database and the
schema, so for example 'cdm_instance.dbo'.}

\item{oracleTempSchema}{A schema where temp tables can be created in Oracle.}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where
the exposure data used to define the exposure cohorts is
available. If exposureTable = DRUG_ERA,
exposureDatabaseSchema is not used but assumed to be
cdmSchema.  Requires read permissions to this database.}

\item{exposureTable}{The tablename that contains the exposure cohorts.  If
exposureTable <> drug_era, then expectation is
exposureTable has format of COHORT table:
cohort_definition_id, subject_id, cohort_start_date,
cohort_end_date.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where
the data used to define the outcome cohorts is available.
If outcomeTable = CONDITION_ERA, outcomeDatabaseSchema is
not used.  Requires read permissions to this database.}

\item{outcomeTable}{The tablename that contains the outcome cohorts.  If
outcomeTable is not CONDITION_OCCURRENCE or CONDITION_ERA,
then expectation is outcomeTable has format of COHORT
table: COHORT_DEFINITION_ID, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{nestingCohortDatabaseSchema}{The name of the database schema that is the location where
the nesting cohort is defined.}

\item{nestingCohortTable}{Name of the table holding the nesting cohort. This table
should have the same structure as the cohort table.}

\item{outputFolder}{Name of the folder where all the outputs will written to.}

\item{ccAnalysisList}{A list of objects of type \code{ccAnalysis} as created
using the \code{\link{createCcAnalysis}} function.}

\item{exposureOutcomeNestingCohortList}{A list of objects of type
\code{exposureOutcomeNestingCohort} as created using the
\code{\link{createExposureOutcomeNestingCohort}} function.}

\item{prefetchExposureData}{Should exposure data for the entire nesting cohort be fetched at
the beginning, or should exposure data be fetch later specifically
for a set of cases and controls. Prefetching can be faster
when there are many outcomes but only few exposures. Prefetching
does not speed up performance when covariates also need to be
constructed.}

\item{compressCaseDataFiles}{Should compression be used when saving?}

\item{getDbCaseDataThreads}{The number of parallel threads to use for building the
caseData objects.}

\item{selectControlsThreads}{The number of parallel threads to use for selecting
controls.}

\item{getDbExposureDataThreads}{The number of parallel threads to use for fetchign data on
exposures for cases and controls.}

\item{createCaseControlDataThreads}{The number of parallel threads to use for creating case
and control data including exposure status indicators}

\item{fitCaseControlModelThreads}{The number of parallel threads to use for fitting the
models.}

\item{cvThreads}{The number of parallel threads used for the
cross-validation to determine the hyper-parameter when
fitting the model.}
}
\description{
Run a list of analyses
}
\details{
Run a list of analyses for the exposure-outcome-nesting cohorts of interest. This function will run
all specified analyses against all hypotheses of interest, meaning that the total number of outcome
models is `length(ccAnalysisList) * length(exposureOutcomeNestingCohortList)` (if all analyses
specify an outcome model should be fitted). When you provide several analyses it will determine
whether any of the analyses have anything in common, and will take advantage of this fact. For
example, if we specify several analyses that only differ in the way the outcome model is fitted,
then this function will extract the data and fit the propensity model only once, and re-use this in
all the analysis.
}
