% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExposureData.R
\name{saveCaseControlsExposure}
\alias{saveCaseControlsExposure}
\title{Save the caseControlsExposure data to folder}
\usage{
saveCaseControlsExposure(caseControlsExposure, folder)
}
\arguments{
\item{caseControlsExposure}{An object of type \code{caseControlsExposure} as generated using
\code{\link{getDbExposureData}}.}

\item{folder}{The name of the folder where the data will be written. The folder
should not yet exist.}
}
\description{
\code{saveCaseControlsExposure} saves an object of type caseControlsExposure to folder.
}
\details{
The data will be written to a set of files in the specified folder.
}
