% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaseData.R
\name{saveCaseData}
\alias{saveCaseData}
\title{Save the case data to file}
\usage{
saveCaseData(caseData, file)
}
\arguments{
\item{caseData}{An object of type \code{CaseData} as generated using
\code{\link{getDbCaseData}}.}

\item{file}{The name of the fle where the data will be written. If the file
exists it will be overwritten.}
}
\description{
\code{saveCaseData} saves an object of type caseData to file
}
\details{
The data will be written to the file specified by the user.
}
