% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{saveCcAnalysisList}
\alias{saveCcAnalysisList}
\title{Save a list of ccAnalysis to file}
\usage{
saveCcAnalysisList(ccAnalysisList, file)
}
\arguments{
\item{ccAnalysisList}{The ccAnalysis list to be written to file}

\item{file}{The name of the file where the results will be written}
}
\description{
Write a list of objects of type \code{ccAnalysis} to file. The file is in JSON format.
}

