% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{saveExposureOutcomeNestingCohortList}
\alias{saveExposureOutcomeNestingCohortList}
\title{Save a list of drugComparatorOutcome to file}
\usage{
saveExposureOutcomeNestingCohortList(exposureOutcomeNestingCohortList,
  file)
}
\arguments{
\item{exposureOutcomeNestingCohortList}{The exposureOutcomeNestingCohort list to be written to
file}

\item{file}{The name of the file where the results will be written}
}
\description{
Write a list of objects of type \code{exposureOutcomeNestingCohort} to file. The file is in JSON
format.
}
