% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectControls.R
\name{selectControls}
\alias{selectControls}
\title{Select matched controls per case}
\usage{
selectControls(caseData, outcomeId, firstOutcomeOnly = TRUE,
  washoutPeriod = 180, controlsPerCase = 2, matchOnAge = TRUE,
  ageCaliper = 2, matchOnGender = TRUE, matchOnProvider = FALSE,
  matchOnCareSite = FALSE, matchOnVisitDate = FALSE,
  visitDateCaliper = 30, matchOnTimeInCohort = FALSE,
  daysInCohortCaliper = 30, minAge = NULL, maxAge = NULL,
  removedUnmatchedCases = TRUE)
}
\arguments{
\item{caseData}{An object of type \code{caseData} as generated using the
\code{\link{getDbCaseData}} function.}

\item{outcomeId}{The outcome ID of the cases for which we need to pick controls.}

\item{firstOutcomeOnly}{Use the first outcome per person?}

\item{washoutPeriod}{Minimum required numbers of days of observation for inclusion as
either case or control.}

\item{controlsPerCase}{Maximum number of controls to select per case.}

\item{matchOnAge}{Match on age?}

\item{ageCaliper}{Maximum difference (in years) in age when matching on age.}

\item{matchOnGender}{Match on gender?}

\item{matchOnProvider}{Match on provider (as specified in the person table)?}

\item{matchOnCareSite}{Match on care site (as specified in the person table)?}

\item{matchOnVisitDate}{Should the index date of the control be changed to the nearest visit
date?}

\item{visitDateCaliper}{Maximum difference (in days) between the index date and the visit
date when matching on visit date.}

\item{matchOnTimeInCohort}{Match on time in nesting cohort? When not using nesting, this is
interpreted as time observed prior to index.}

\item{daysInCohortCaliper}{Maximum difference (in days) in time in cohort.}

\item{minAge}{Minimum age at which patient time will be included in the analysis.
Note that information prior to the min age is still used to determine
exposure status after the minimum age (e.g. when a prescription was
started just prior to reaching the minimum age). Also, outcomes
occurring before the minimum age is reached will be considered as
prior outcomes when using first outcomes only. Age should be specified
in years, but non-integer values are allowed. If not specified, no age
restriction will be applied.}

\item{maxAge}{Maximum age at which patient time will be included in the analysis. Age
should be specified in years, but non-integer values are allowed. If not
specified, no age restriction will be applied.}

\item{removedUnmatchedCases}{Should cases with no matched controls be removed?}
}
\value{
A data frame with these columns: \describe{ \item{personId}{The person ID} \item{indexDate}{The
index date} \item{isCase}{Is the person a case or a control?} \item{stratumId}{The ID linking cases
and controls in a matched set} }
}
\description{
Select matched controls per case
}
\details{
Select controls per case. Controls are matched on calendar time and the criteria defined in the
arguments. Controls are randomly sampled to the required number.
}

