% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createGetDbCaseDataArgs}
\alias{createGetDbCaseDataArgs}
\title{Create a parameter object for the function getDbCaseData}
\usage{
createGetDbCaseDataArgs(useNestingCohort = FALSE,
  useObservationEndAsNestingEndDate = TRUE, getVisits = TRUE,
  studyStartDate = "", studyEndDate = "")
}
\arguments{
\item{useNestingCohort}{Should the study be nested in a cohort (e.g. people witha
specific indication)? If not, the study will be nestedin
the general population.}

\item{useObservationEndAsNestingEndDate}{When using a nesting cohort, should the observationperiod
end date be used instead of the cohort end date?}

\item{getVisits}{Get data on visits? This is needed when matching on
visitdate is requested later on.}

\item{studyStartDate}{A calendar date specifying the minimum date where data
isused. Date format is 'yyyymmdd'.}

\item{studyEndDate}{A calendar date specifying the maximum date where data
isused. Date format is 'yyyymmdd'.}
}
\description{
Create a parameter object for the function getDbCaseData
}
\details{
Create an object defining the parameter values.
}
