% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createGetDbExposureDataArgs}
\alias{createGetDbExposureDataArgs}
\title{Create a parameter object for the function getDbExposureData}
\usage{
createGetDbExposureDataArgs(covariateSettings = NULL)
}
\arguments{
\item{covariateSettings}{An object of type covariateSettings as created using
thecreateCovariateSettings function in theFeatureExtraction package. If
NULL then no covariate data isretrieved.}
}
\description{
Create a parameter object for the function getDbExposureData
}
\details{
Create an object defining the parameter values.
}

