% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFitting.R
\name{fitCaseControlModel}
\alias{fitCaseControlModel}
\title{Fit the case-control model}
\usage{
fitCaseControlModel(
  caseControlData,
  useCovariates = FALSE,
  excludeCovariateIds = c(),
  includeCovariateIds = c(),
  caseControlsExposure = NULL,
  profileGrid = seq(log(0.1), log(10), length.out = 1000),
  prior = createPrior("laplace", useCrossValidation = TRUE),
  control = createControl(cvType = "auto", startingVariance = 0.01, tolerance = 2e-07,
    cvRepetitions = 10, selectorType = "byPid", noiseLevel = "quiet")
)
}
\arguments{
\item{caseControlData}{A data frame as generated by the \code{\link{createCaseControlData}}
function.}

\item{useCovariates}{Whether to use the covariates in the \code{caseControlsExposure}.}

\item{excludeCovariateIds}{Exclude these covariates from the model.}

\item{includeCovariateIds}{Include only these covariates in the model.}

\item{caseControlsExposure}{An object of type \code{caseControlsExposure} as created using the
\code{\link{getDbExposureData}} function.}

\item{profileGrid}{A one-dimensional grid of points on the log(relative risk) scale where
the likelihood for the treatment variable coefficient is sampled. Set to
NULL to skip profiling.}

\item{prior}{The prior used to fit the model. See
\code{\link[Cyclops]{createPrior}} for details.}

\item{control}{The control object used to control the cross-validation used to
determine the hyperparameters of the prior (if applicable). See
\code{\link[Cyclops]{createControl}} for details.}
}
\value{
An object of type \code{outcomeModel}.
}
\description{
Fit the case-control model
}
\details{
Fits the model using a conditional logistic regression.
}
