% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFitting.R
\name{fitCaseControlModel}
\alias{fitCaseControlModel}
\title{Fit the case-control model}
\usage{
fitCaseControlModel(caseControlData, useCovariates = FALSE,
  excludeCovariateIds = c(), includeCovariateIds = c(),
  caseControlsExposure = NULL, prior = createPrior("laplace",
  useCrossValidation = TRUE), control = createControl(cvType = "auto",
  startingVariance = 0.01, tolerance = 2e-07, cvRepetitions = 10, selectorType =
  "byPid", noiseLevel = "quiet"))
}
\arguments{
\item{caseControlData}{A data frame as generated by the \code{\link{createCaseControlData}} function.}
}
\value{
An object of type \code{outcomeModel}.
}
\description{
Fit the case-control model
}
\details{
Fits the model using a conditional logistic regression.
}

