% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExposureData.R
\name{getDbExposureData}
\alias{getDbExposureData}
\title{Get exposure data for cases and controls from a database}
\usage{
getDbExposureData(caseControls, connectionDetails, oracleTempSchema = NULL,
  exposureDatabaseSchema, exposureTable = "drug_era", exposureIds = c())
}
\arguments{
\item{caseControls}{A data frame as generated by the \code{\link{selectControls}} function.}

\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{oracleTempSchema}{A schema where temp tables can be created in Oracle.}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where
the exposure data used to define the exposure cohorts is
available. If exposureTable = DRUG_ERA,
exposureDatabaseSchema is not used but assumed to be
cdmSchema.  Requires read permissions to this database.}

\item{exposureTable}{The tablename that contains the exposure cohorts.  If
exposureTable <> drug_era, then expectation is exposureTable
has format of COHORT table: cohort_definition_id, subject_id,
cohort_start_date, cohort_end_date.}

\item{exposureIds}{A list of identifiers to define the exposures of interest. If
exposureTable = drug_era, exposureIds should be concept_id.
If exposureTable <> drug_era, exposureIds is used to select
the cohort_definition_id in the cohort-like table. If no
exposureIds are provided, all drugs or cohorts in the
exposureTable are included as exposures.}
}
\description{
Get exposure data for cases and controls from a database
}

