% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExposureData.R
\name{getDbExposureData}
\alias{getDbExposureData}
\title{Get exposure data for cases and controls from a database}
\usage{
getDbExposureData(
  caseControls,
  connectionDetails,
  oracleTempSchema = NULL,
  exposureDatabaseSchema = NULL,
  exposureTable = "drug_era",
  exposureIds = c(),
  cdmDatabaseSchema = exposureDatabaseSchema,
  covariateSettings = NULL,
  caseData = NULL
)
}
\arguments{
\item{caseControls}{A data frame as generated by the \code{\link{selectControls}}
function.}

\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{oracleTempSchema}{A schema where temp tables can be created in Oracle.}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where the
exposure data used to define the exposure cohorts is available. If
exposureTable = DRUG_ERA, exposureDatabaseSchema is not used but
assumed to be cdmSchema.  Requires read permissions to this
database.}

\item{exposureTable}{The tablename that contains the exposure cohorts.  If exposureTable
<> drug_era, then expectation is exposureTable has format of COHORT
table: cohort_definition_id, subject_id, cohort_start_date,
cohort_end_date.}

\item{exposureIds}{A list of identifiers to define the exposures of interest. If
exposureTable = drug_era, exposureIds should be concept_id. If
exposureTable <> drug_era, exposureIds is used to select the
cohort_definition_id in the cohort-like table. If no exposureIds are
provided, all drugs or cohorts in the exposureTable are included as
exposures.}

\item{cdmDatabaseSchema}{Needed when constructing covariates: the name of the database schema
that contains the OMOP CDM instance.  Requires read permissions to
this database. On SQL Server, this should specify both the database
and the schema, so for example 'cdm_instance.dbo'.}

\item{covariateSettings}{An object of type \code{covariateSettings} as created using the
\code{createCovariateSettings} function in the
\code{FeatureExtraction} package. If NULL then no covariate data is
retrieved.}

\item{caseData}{An object of type \code{caseData} as generated using the
\code{\link{getDbCaseData}} function. If caseData is provided and contains
the exposure data (see getExposures in the  \code{\link{getDbCaseData}} function,
and if no covariates need to constructed (covariateSettings = NULL), then the
no connection to the database is used to create the exposure data. This may be
much more efficient in some situations.}
}
\description{
If additional exposure data or covariate data is required, this function will send the information
on the cases and controls back to the server.

Note: For PDW and RedShift, where uploading data can be excrutiatingly slow, you can use bulk loading
by preparing the environment as described in the \code{\link[DatabaseConnector]{insertTable}} function
in the DatabaseConnection package, and setting \code{oracleTempSchema} to a schema where you have write
privileges (bulk import can only upload to permanent tables).
}
