% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadingSaving.R
\name{loadCaseData}
\alias{loadCaseData}
\title{Load the case data from a folder}
\usage{
loadCaseData(folder, readOnly = TRUE)
}
\arguments{
\item{folder}{The name of the folder containing the data.}

\item{readOnly}{If true, the data is opened read only.}
}
\value{
An object of class \code{caseData}.
}
\description{
\code{loadCaseData} loads an object of type caseData from a folder in the file system.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
