% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadingSaving.R
\name{saveCaseData}
\alias{saveCaseData}
\title{Save the case data to folder}
\usage{
saveCaseData(caseData, folder)
}
\arguments{
\item{caseData}{An object of type \code{caseData} as generated using \code{\link{getDbCaseData}}.}

\item{folder}{The name of the folder where the data will be written. The folder should not yet
exist.}
}
\description{
\code{saveCaseData} saves an object of type caseData to folder.
}
\details{
The data will be written to a set of files in the specified folder.
}

