% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectControls.R
\name{selectControls}
\alias{selectControls}
\title{Select controls}
\usage{
selectControls(caseData, outcomeId, firstOutcomeOnly = TRUE,
  washoutPeriod = 180,
  controlSelectionCriteria = createMatchingCriteria(), minAge = NULL,
  maxAge = NULL)
}
\arguments{
\item{caseData}{An object of type \code{caseData} as generated using the
\code{\link{getDbCaseData}} function.}

\item{outcomeId}{The outcome ID of the cases for which we need to pick controls.}

\item{firstOutcomeOnly}{Use the first outcome per person?}

\item{washoutPeriod}{Minimum required numbers of days of observation for inclusion as
either case or control.}

\item{controlSelectionCriteria}{Either a \code{matchingCriteria} object as generated using the \code{
\link{createMatchingCriteria}} function, or a \code{samplingCriteria}
object as generated using the \code{link{createSamplingCriteria}}
function.}

\item{minAge}{Minimum age at which patient time will be included in the analysis.
Note that information prior to the min age is still used to determine
exposure status after the minimum age (e.g. when a prescription was
started just prior to reaching the minimum age). Also, outcomes
occurring before the minimum age is reached will be considered as
prior outcomes when using first outcomes only. Age should be specified
in years, but non-integer values are allowed. If not specified, no age
restriction will be applied.}

\item{maxAge}{Maximum age at which patient time will be included in the analysis. Age
should be specified in years, but non-integer values are allowed. If not
specified, no age restriction will be applied.}
}
\value{
A data frame with these columns: \describe{ \item{personId}{The person ID} \item{indexDate}{The
index date} \item{isCase}{Is the person a case or a control?} \item{stratumId}{The ID linking cases
and controls in a matched set (only available when matching)} }
}
\description{
Select controls
}
\details{
Select controls either by individually matching controls to each case, or by random sampling controls from the
(nested) population.
}
